test: add create_comment integration test module
